src/geo-location-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
requests==2.31.0
pymongo==4.6.1
httpx==0.26.0